Correctly update endpoints in secant algorithm
This is kind of funny--I updated `lo` and `hi` but not `lo_volume`
and `hi_volume`. It's kind of surprising that that the algorithm
worked at all, but if you think about it, in the broken
implementation `frac` was a constant between 0 and 1, so the
algorithm was essentially equivalent to bisection using a value
other than 0.5.